Test: Separate out settings override
