+++ test/test_library.py
@@ -1,7 +1,7 @@
 import os
 import sys
 
-from pydantic_settings import SettingsConfigDict
+from conftest import settings_override
 
 # getting the name of the directory
 # where the this file is present.
@@ -16,62 +16,11 @@ parent = os.path.dirname(current)
 sys.path.append(parent)
 
 from src.library import combine_library_lists
-from src.settings import AppSettings
-
-
-class _IsolatedAppSettings(AppSettings):
-    """
-    AppSettings that ignores all external configuration sources (env vars,
-    .env, legacy .env, config.yaml) so tests depend only on the kwargs passed
-    in. Without this, constructing AppSettings would read the developer's real
-    .env / config.yaml and contaminate the test.
-    """
-
-    model_config = SettingsConfigDict(
-        yaml_file=None,
-        yaml_file_encoding=None,
-        nested_model_default_partial_update=True,
-        extra="forbid",
-    )
-
-    @classmethod
-    def settings_customise_sources(
-        cls,
-        settings_cls,
-        init_settings,
-        env_settings,
-        dotenv_settings,
-        file_secret_settings,
-    ):
-        return (init_settings,)
-
-
-def _settings(**overrides) -> AppSettings:
-    base = {
-        "plex": [
-            {
-                "name": "plex-main",
-                "baseurl": "http://plex",
-                "token": "x",
-                "sync_to": ["jellyfin-main"],
-            }
-        ],
-        "jellyfin": [
-            {
-                "name": "jellyfin-main",
-                "baseurl": "http://jellyfin",
-                "token": "x",
-                "sync_to": ["plex-main"],
-            }
-        ],
-    }
-    base.update(overrides)
-    return _IsolatedAppSettings(**base)
 
 
 def test_combine_library_lists_implicit_same_name():
     """Libraries with the same name on both servers sync without a mapping."""
-    settings = _settings()
+    settings = settings_override()
 
     # {library_name: type}
     server_1_libs = {"Movies": "movie", "TV Shows": "show"}
@@ -88,7 +37,7 @@ def test_combine_library_lists_implicit_same_name():
 
 def test_combine_library_lists_with_mapping():
     """A library_mappings entry links differently-named libraries."""
-    settings = _settings(
+    settings = settings_override(
         library_mappings=[
             {
                 "canonical": "Shows",
@@ -112,7 +61,7 @@ def test_combine_library_lists_with_mapping():
 
 def test_combine_library_lists_type_blacklist():
     """A library whose type is blacklisted is dropped regardless of name match."""
-    settings = _settings(blacklist_library_types=["music"])
+    settings = settings_override(blacklist_library_types=["music"])
 
     server_1_libs = {"Movies": "movie", "Music": "music"}
     server_2_libs = {"Movies": "movies", "Music": "music"}
@@ -127,7 +76,7 @@ def test_combine_library_lists_type_blacklist():
 
 def test_combine_library_lists_type_whitelist():
     """With a type whitelist set, only matching types sync."""
-    settings = _settings(whitelist_library_types=["movie", "movies"])
+    settings = settings_override(whitelist_library_types=["movie", "movies"])
 
     server_1_libs = {"Movies": "movie", "TV Shows": "show"}
     server_2_libs = {"Movies": "movies", "Shows": "tvshows"}
@@ -141,7 +90,7 @@ def test_combine_library_lists_type_whitelist():
 
 def test_combine_library_lists_name_blacklist():
     """A blacklisted library name is filtered out by should_sync_library."""
-    settings = _settings(blacklist_libraries=["TV Shows"])
+    settings = settings_override(blacklist_libraries=["TV Shows"])
 
     server_1_libs = {"Movies": "movie", "TV Shows": "show"}
     server_2_libs = {"Movies": "movies", "TV Shows": "tvshows"}
@@ -155,7 +104,7 @@ def test_combine_library_lists_name_blacklist():
 
 def test_combine_library_lists_name_whitelist():
     """With a library whitelist set, only whitelisted libraries sync."""
-    settings = _settings(whitelist_libraries=["Movies"])
+    settings = settings_override(whitelist_libraries=["Movies"])
 
     server_1_libs = {"Movies": "movie", "TV Shows": "show"}
     server_2_libs = {"Movies": "movies", "TV Shows": "tvshows"}
