+++ test/test_watched.py
@@ -2,7 +2,7 @@ import os
 import sys
 from datetime import datetime
 
-from pydantic_settings import SettingsConfigDict
+from conftest import settings_override
 
 # getting the name of the directory
 # where the this file is present.
@@ -16,8 +16,6 @@ parent = os.path.dirname(current)
 # the sys.path.
 sys.path.append(parent)
 
-
-from src.settings import AppSettings
 from src.watched import (
     LibraryData,
     MediaIdentifiers,
@@ -27,62 +25,6 @@ from src.watched import (
     WatchedStatus,
     cleanup_watched,
 )
-
-
-class _IsolatedAppSettings(AppSettings):
-    """
-    AppSettings that ignores all external configuration sources (env vars,
-    .env, legacy .env, config.yaml) so tests depend only on the kwargs passed
-    in. Without this, constructing AppSettings would read the developer's real
-    .env / config.yaml, which can both contaminate results and raise
-    validation errors when the ambient config conflicts with the test config.
-    """
-
-    model_config = SettingsConfigDict(
-        yaml_file=None,
-        yaml_file_encoding=None,
-        nested_model_default_partial_update=True,
-        extra="forbid",
-    )
-
-    @classmethod
-    def settings_customise_sources(
-        cls,
-        settings_cls,
-        init_settings,
-        env_settings,
-        dotenv_settings,
-        file_secret_settings,
-    ):
-        return (init_settings,)
-
-
-def _settings() -> AppSettings:
-    """
-    A minimal two-server AppSettings. No user_mappings / library_mappings are
-    declared, so cleanup_watched resolves users and libraries via the implicit
-    same-name fallback — matching how these test fixtures key everything by
-    identical names ('user1', 'TV Shows', 'Movies', ...).
-    """
-    return _IsolatedAppSettings(
-        plex=[
-            {
-                "name": "server1",
-                "baseurl": "http://server1",
-                "token": "x",
-                "sync_to": ["server2"],
-            }
-        ],
-        jellyfin=[
-            {
-                "name": "server2",
-                "baseurl": "http://server2",
-                "token": "x",
-                "sync_to": ["server1"],
-            }
-        ],
-    )
-
 
 viewed_date = datetime.today()
 
@@ -649,7 +591,7 @@ tv_shows_2_watched_list_1: list[Series] = [
 
 
 def test_simple_cleanup_watched():
-    settings = _settings()
+    settings = settings_override()
     server_1 = "server1"
     server_2 = "server2"
     average_time = 0.0
